Refactor code - 리뷰 반영(패키지 구조 변경, 테스트 분리)
