+++ app/src/main/java/kr/flab/movieon/integrate/security/domain/JpaRefreshTokenInfoRepository.java
@@ -1,0 +1,8 @@
+package kr.flab.movieon.integrate.security.domain;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface JpaRefreshTokenInfoRepository extends RefreshTokenInfoRepository,
+    JpaRepository<RefreshTokenInfo, Long> {
+
+}
