+++ app/src/main/java/kr/flab/movieon/integrate/security/domain/RefreshTokenInfoRepository.java
@@ -1,0 +1,14 @@
+package kr.flab.movieon.integrate.security.domain;
+
+import java.util.Optional;
+
+public interface RefreshTokenInfoRepository {
+
+    boolean existsByRefreshTokenJti(String refreshTokenJti);
+
+    Optional<RefreshTokenInfo> findByRefreshTokenJti(String refreshTokenJti);
+
+    RefreshTokenInfo save(RefreshTokenInfo entity);
+
+    void delete(RefreshTokenInfo entity);
+}
