+++ app/src/main/java/kr/flab/movieon/integrate/security/domain/Token.java
@@ -1,0 +1,6 @@
+package kr.flab.movieon.integrate.security.domain;
+
+public interface Token {
+
+    String getToken();
+}
