+++ app/src/main/java/kr/flab/movieon/integrate/security/domain/TokenConverter.java
@@ -1,0 +1,10 @@
+package kr.flab.movieon.integrate.security.domain;
+
+/**
+ * Convert Refresh Token to Access Token.
+ * If the conversion succeeds, the existing Refresh Token is expired.
+ */
+public interface TokenConverter {
+
+    Token convert(String payload);
+}
