+++ app/src/main/java/kr/flab/movieon/integrate/security/domain/TokenExtractor.java
@@ -1,0 +1,6 @@
+package kr.flab.movieon.integrate.security.domain;
+
+public interface TokenExtractor {
+
+    String extract(String payload);
+}
