+++ app/src/main/java/kr/flab/movieon/integrate/security/domain/TokenGenerator.java
@@ -1,0 +1,8 @@
+package kr.flab.movieon.integrate.security.domain;
+
+public interface TokenGenerator {
+
+    Token createRefreshToken(AccountContext accountContext);
+
+    Token createAccessToken(AccountContext accountContext);
+}
