+++ app/src/main/java/kr/flab/movieon/integrate/security/domain/TokenVerifier.java
@@ -1,0 +1,6 @@
+package kr.flab.movieon.integrate.security.domain;
+
+public interface TokenVerifier {
+
+    boolean verify(String jti);
+}
