+++ app/src/main/java/kr/flab/movieon/integrate/security/exception/JwtExpiredTokenException.java
@@ -1,0 +1,10 @@
+package kr.flab.movieon.integrate.security.exception;
+
+import org.springframework.security.core.AuthenticationException;
+
+public final class JwtExpiredTokenException extends AuthenticationException {
+
+    public JwtExpiredTokenException(String message) {
+        super(message);
+    }
+}
