+++ app/src/main/java/kr/flab/movieon/integrate/security/jwt/JwtToken.java
@@ -1,0 +1,27 @@
+package kr.flab.movieon.integrate.security.jwt;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import io.jsonwebtoken.Claims;
+import kr.flab.movieon.integrate.security.domain.Token;
+
+public final class JwtToken implements Token {
+
+    private final String token;
+
+    @JsonIgnore
+    private final Claims claims;
+
+    public JwtToken(String token, Claims claims) {
+        this.token = token;
+        this.claims = claims;
+    }
+
+    @Override
+    public String getToken() {
+        return this.token;
+    }
+
+    public Claims getClaims() {
+        return claims;
+    }
+}
