+++ app/src/main/java/kr/flab/movieon/integrate/security/jwt/Scopes.java
@@ -1,0 +1,5 @@
+package kr.flab.movieon.integrate.security.jwt;
+
+public enum Scopes {
+    REFRESH_TOKEN
+}
