--- app/src/main/java/kr/flab/movieon/security/domain/AuthEntityRepository.java
@@ -1,14 +1,0 @@
-package kr.flab.movieon.security.domain;
-
-import java.util.Optional;
-
-public interface AuthEntityRepository {
-
-    boolean existsByRefreshTokenJti(String refreshTokenJti);
-
-    Optional<AuthEntity> findByRefreshTokenJti(String refreshTokenJti);
-
-    AuthEntity save(AuthEntity entity);
-
-    void delete(AuthEntity entity);
-}
