--- app/src/main/java/kr/flab/movieon/security/domain/JpaAuthEntityRepository.java
@@ -1,8 +1,0 @@
-package kr.flab.movieon.security.domain;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface JpaAuthEntityRepository extends AuthEntityRepository,
-    JpaRepository<AuthEntity, Long> {
-
-}
