--- app/src/main/java/kr/flab/movieon/security/domain/Token.java
@@ -1,6 +1,0 @@
-package kr.flab.movieon.security.domain;
-
-public interface Token {
-
-    String getToken();
-}
