--- app/src/main/java/kr/flab/movieon/security/domain/TokenConverter.java
@@ -1,10 +1,0 @@
-package kr.flab.movieon.security.domain;
-
-/**
- * Convert Refresh Token to Access Token.
- * If the conversion succeeds, the existing Refresh Token is expired.
- */
-public interface TokenConverter {
-
-    Token convert(String payload);
-}
