--- app/src/main/java/kr/flab/movieon/security/domain/TokenExtractor.java
@@ -1,6 +1,0 @@
-package kr.flab.movieon.security.domain;
-
-public interface TokenExtractor {
-
-    String extract(String payload);
-}
