--- app/src/main/java/kr/flab/movieon/security/domain/TokenGenerator.java
@@ -1,8 +1,0 @@
-package kr.flab.movieon.security.domain;
-
-public interface TokenGenerator {
-
-    Token createRefreshToken(AccountContext accountContext);
-
-    Token createAccessToken(AccountContext accountContext);
-}
