--- app/src/main/java/kr/flab/movieon/security/domain/TokenVerifier.java
@@ -1,6 +1,0 @@
-package kr.flab.movieon.security.domain;
-
-public interface TokenVerifier {
-
-    boolean verify(String jti);
-}
