--- app/src/main/java/kr/flab/movieon/security/exception/InvalidTokenException.java
@@ -1,15 +1,0 @@
-package kr.flab.movieon.security.exception;
-
-import kr.flab.movieon.account.domain.exception.SystemException;
-import kr.flab.movieon.common.ErrorCode;
-
-public final class InvalidTokenException extends SystemException {
-
-    public InvalidTokenException(String message) {
-        super(message, ErrorCode.INVALID_TOKEN);
-    }
-
-    public InvalidTokenException() {
-        super(ErrorCode.INVALID_TOKEN);
-    }
-}
