--- app/src/main/java/kr/flab/movieon/security/jwt/Scopes.java
@@ -1,5 +1,0 @@
-package kr.flab.movieon.security.jwt;
-
-public enum Scopes {
-    REFRESH_TOKEN
-}
