+++ app/src/test/java/kr/flab/movieon/security/domain/TokenConvertTest.java
@@ -12,12 +12,14 @@ import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.atomic.AtomicLong;
 import kr.flab.movieon.account.domain.Account;
 import kr.flab.movieon.account.domain.AccountRepository;
-import kr.flab.movieon.security.SecurityAppProperties;
-import kr.flab.movieon.security.exception.InvalidTokenException;
-import kr.flab.movieon.security.jwt.JwtTokenConverter;
-import kr.flab.movieon.security.jwt.JwtTokenExtractor;
-import kr.flab.movieon.security.jwt.JwtTokenGenerator;
-import kr.flab.movieon.security.jwt.JwtTokenVerifier;
+import kr.flab.movieon.integrate.security.SecurityAppProperties;
+import kr.flab.movieon.integrate.security.domain.RefreshTokenInfo;
+import kr.flab.movieon.integrate.security.domain.RefreshTokenInfoRepository;
+import kr.flab.movieon.integrate.security.exception.InvalidTokenException;
+import kr.flab.movieon.integrate.security.jwt.JwtTokenConverter;
+import kr.flab.movieon.integrate.security.jwt.JwtTokenExtractor;
+import kr.flab.movieon.integrate.security.jwt.JwtTokenGenerator;
+import kr.flab.movieon.integrate.security.jwt.JwtTokenVerifier;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.DynamicTest;
 import org.junit.jupiter.api.TestFactory;
@@ -25,55 +27,30 @@ import org.springframework.security.authentication.AuthenticationServiceExceptio
 
 final class TokenConvertTest {
 
-    private final SecurityAppProperties securityAppProperties = new SecurityAppProperties(1800,
-        "movieon-test", "c2VjcmV0S2V5LXRlc3QtYXV0aG9yaXphdGlvbi1qd3QtbWFuYWdlLXRva2Vu", 604800);
+    private final String invalidToken = "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
+        + ".eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ"
+        + ".-VehURC5AsJrkeIUxxUzr5_kSHnyuEqn63XjTGqVCIk";
+    private final SecurityAppProperties securityAppProperties = new SecurityAppProperties(
+        1800,
+        "movieon-test",
+        "c2VjcmV0S2V5LXRlc3QtYXV0aG9yaXphdGlvbi1qd3QtbWFuYWdlLXRva2Vu",
+        604800);
 
     @TestFactory
     @DisplayName("JWT 토큰 발행 및 변환 테스트")
-    Collection<DynamicTest> token_publish_and_convert_scenarios() {
-        var authRepository = new FakeAuthEntityRepository();
-        var tokenGenerator = new JwtTokenGenerator(authRepository, securityAppProperties);
-        var tokenVerifier = new JwtTokenVerifier(authRepository);
-        var tokenExtractor = new JwtTokenExtractor();
-        var accountRepository = new StubAccountRepository();
+    Collection<DynamicTest> token_converter_test() {
+        var authRepository = new FakeRefreshTokenInfoRepository();
 
-        var tokenConverter = new JwtTokenConverter(accountRepository, authRepository,
-            securityAppProperties, tokenGenerator, tokenVerifier, tokenExtractor);
-
-        var account = accountRepository.getAccountStub();
-        var invalidToken = "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9"
-            + ".eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ"
-            + ".-VehURC5AsJrkeIUxxUzr5_kSHnyuEqn63XjTGqVCIk";
+        var tokenConverter = new JwtTokenConverter(
+            new StubAccountRepository(),
+            authRepository,
+            securityAppProperties,
+            new JwtTokenGenerator(authRepository, securityAppProperties),
+            new JwtTokenVerifier(authRepository),
+            new JwtTokenExtractor());
 
         return Arrays.asList(
-            dynamicTest("공백인 payload 에서 토큰을 추출하려고 하면 인증 예외가 발생한다.", () -> {
-                assertThatExceptionOfType(AuthenticationServiceException.class).isThrownBy(
-                    () -> tokenExtractor.extract(""));
-
-            }), dynamicTest("null 에서 토큰을 추출하려고 하면 인증 예외가 발생한다.", () -> {
-                assertThatExceptionOfType(AuthenticationServiceException.class).isThrownBy(
-                    () -> tokenExtractor.extract(null));
-
-            }), dynamicTest("유효하지 않은 문자열에서 토큰을 추출하려고 하면 예외가 발생한다.", () -> {
-                assertThatExceptionOfType(AuthenticationServiceException.class).isThrownBy(
-                    () -> tokenExtractor.extract("BeBeBe DERFTGYHUJIKOL"));
-
-            }), dynamicTest("JTI 값이 null 이라면 검증에 실패한다.", () -> {
-                var isVerified = tokenVerifier.verify(null);
-
-                assertThat(isVerified).isFalse();
-
-            }), dynamicTest("JTI 값이 공백 이라면 검증에 실패한다.", () -> {
-                var isVerified = tokenVerifier.verify("");
-
-                assertThat(isVerified).isFalse();
-
-            }), dynamicTest("유효하지 않은 토큰을 변환하면 예외를 반환한다.", () -> {
-                var isVerified = tokenVerifier.verify(invalidToken);
-
-                assertThat(isVerified).isFalse();
-
-            }), dynamicTest("토큰 값이 null 이라면 변환에 인증 예외가 발생한다.", () -> {
+            dynamicTest("토큰 값이 null 이라면 변환에 인증 예외가 발생한다.", () -> {
                 assertThatExceptionOfType(AuthenticationServiceException.class).isThrownBy(
                     () -> tokenConverter.convert(null));
 
@@ -84,9 +61,53 @@ final class TokenConvertTest {
             }));
     }
 
-    private final class FakeAuthEntityRepository implements AuthEntityRepository {
+    @TestFactory
+    @DisplayName("JWT Refresh 토큰 검증 테스트")
+    Collection<DynamicTest> token_verifier_test() {
+        var tokenVerifier = new JwtTokenVerifier(new FakeRefreshTokenInfoRepository());
 
-        private final Map<Long, AuthEntity> data = new ConcurrentHashMap<>();
+        return Arrays.asList(
+            dynamicTest("유효하지 않은 토큰은 검증에 실패한다.", () -> {
+                var isVerified = tokenVerifier.verify(invalidToken);
+
+                assertThat(isVerified).isFalse();
+
+            }), dynamicTest("JTI 값이 null 이라면 검증에 실패한다.", () -> {
+                var isVerified = tokenVerifier.verify(null);
+
+                assertThat(isVerified).isFalse();
+
+            }), dynamicTest("JTI 값이 공백 이라면 검증에 실패한다.", () -> {
+                var isVerified = tokenVerifier.verify("");
+
+                assertThat(isVerified).isFalse();
+            }));
+    }
+
+    @TestFactory
+    @DisplayName("JWT 토큰 추출 테스트")
+    Collection<DynamicTest> token_extractor_test() {
+        var tokenExtractor = new JwtTokenExtractor();
+
+        return Arrays.asList(
+            dynamicTest("공백인 payload 에서 토큰을 추출하려고 하면 인증 예외가 발생한다.", () -> {
+                assertThatExceptionOfType(AuthenticationServiceException.class).isThrownBy(
+                    () -> tokenExtractor.extract(""));
+
+            }), dynamicTest("null 에서 토큰을 추출하려고 하면 인증 예외가 발생한다.", () -> {
+                assertThatExceptionOfType(AuthenticationServiceException.class).isThrownBy(
+                    () -> tokenExtractor.extract(null));
+
+            }), dynamicTest("유효하지 않은 문자열에서 토큰을 추출하려고 하면 인증 예외가 발생한다.", () -> {
+                assertThatExceptionOfType(AuthenticationServiceException.class).isThrownBy(
+                    () -> tokenExtractor.extract("BeBeBe DERFTGYHUJIKOL"));
+
+            }));
+    }
+
+    private final class FakeRefreshTokenInfoRepository implements RefreshTokenInfoRepository {
+
+        private final Map<Long, RefreshTokenInfo> data = new ConcurrentHashMap<>();
         private final AtomicLong idGenerator = new AtomicLong();
 
         @Override
@@ -96,18 +117,18 @@ final class TokenConvertTest {
         }
 
         @Override
-        public Optional<AuthEntity> findByRefreshTokenJti(String refreshTokenJti) {
+        public Optional<RefreshTokenInfo> findByRefreshTokenJti(String refreshTokenJti) {
             return data.values().stream()
                 .filter(entity -> entity.getRefreshTokenJti().equals(refreshTokenJti)).findAny();
         }
 
         @Override
-        public AuthEntity save(AuthEntity entity) {
+        public RefreshTokenInfo save(RefreshTokenInfo entity) {
             return data.put(idGenerator.incrementAndGet(), entity);
         }
 
         @Override
-        public void delete(AuthEntity entity) {
+        public void delete(RefreshTokenInfo entity) {
             data.remove(entity.getId());
         }
     }
